Move common.signal_models to common.models.signal
