+++ hasher-matcher-actioner/hmalib/lambdas/pdq/pdq_matcher.py
@@ -15,7 +15,7 @@ from threatexchange.signal_type.pdq import PdqSignal
 from hmalib import metrics
 from hmalib.models import MatchRecord
 from hmalib.common.message_models import BankedSignal, MatchMessage
-from hmalib.common.signal_models import PDQSignalMetadata
+from hmalib.common.models.signal import PDQSignalMetadata
 from hmalib.common.logging import get_logger
 from hmalib.common.config import HMAConfig
 from hmalib.common.configs.fetcher import ThreatExchangeConfig
